+++ src/main/java/it/polimi/ingsw/model/card/MovementCharacterCard.java
@@ -13,9 +13,9 @@ public class MovementCharacterCard extends CharacterCard {
     public MovementCharacterCard(CharacterCard card) {
         super(card);
         Map<String, Object> params = this.getParams();
-        this.extraControl = (int) params.get("extra_control");
-        this.possibleExchange = (int) params.get("possible_exchange");
-        this.numOfStudentsToReturn = (int) params.get("num_of_students_to_return");
+        this.extraControl = ((Double) params.get("extra_control")).intValue();
+        this.possibleExchange = ((Double) params.get("possible_exchange")).intValue();
+        this.numOfStudentsToReturn = ((Double) params.get("num_of_students_to_return")).intValue();
     }
 
     public int getExtraControl() {
